products table added

+++ Models/Product.cs
@@ -9,7 +9,6 @@ public class Product
     [BsonRepresentation(BsonType.ObjectId)]
     public string id { get; private set; }
     [BsonElement("name")] public string Name { get; set; }
-
-    [BsonElement("price")] public double Price { get; set; }
+    [BsonElement("price")] public int Price { get; set; }
     [BsonElement("maxQuantity")] public int MaxQuantity { get;  set; }
 }
